% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{p_coding}
\alias{p_coding}
\title{Format p-values consistently}
\usage{
p_coding(pval, digits = 3)
}
\arguments{
\item{pval}{Numeric p-value.}

\item{digits}{Rounding digits when shown numerically.}
}
\value{
Character.
}
\description{
Format p-values consistently
}
\keyword{internal}
