% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{round_tidy_half_up}
\alias{round_tidy_half_up}
\title{Half-up rounding with tidy trailing zeros}
\usage{
round_tidy_half_up(x, digits = 3)
}
\arguments{
\item{x}{numeric}

\item{digits}{integer number of decimals}
}
\value{
character vector with fixed decimals
}
\description{
Half-up rounding with tidy trailing zeros
}
\keyword{internal}
