% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\encoding{UTF-8}
\name{tee_faktorianalyysitaulukko}
\alias{tee_faktorianalyysitaulukko}
\title{Faktorianalyysin tulostaulukko}
\usage{
tee_faktorianalyysitaulukko(
  faktorimalli,
  kieli = "suomi",
  desimaalierotin = "piste",
  piilota_lataukset_alle = 0.3
)
}
\arguments{
\item{faktorimalli}{faktorianalyysimalli}

\item{kieli}{Jos suomi, kääntää käsitteet suomeksi, muuten jättää englanninkieliset.}

\item{desimaalierotin}{Jos pilkku, muuttaa valmiin taulukon desimaalierottimen pilkuksi, muuten jättää alkuperäisen pisteen.}

\item{piilota_lataukset_alle}{Poistaa taulukosta argumentin itseisarvoa pienemmät lataukset.}
}
\value{
Valmis faktorianalyysitaulukko.
}
\description{
Funktio tekee psych-paketin faktorianalyysimallista HY:n sosiaalitieteiden koulutusohjelmien ohjeiden mukaisen tulostaulukon.
}
