% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_jatkuvan_muuttujan_esittely}
\alias{tee_jatkuvan_muuttujan_esittely}
\title{Esitellään jatkuva muuttuja.}
\usage{
tee_jatkuvan_muuttujan_esittely(aineisto, ..., lv = F, vinous = F)
}
\arguments{
\item{aineisto}{Aineistotiedosto.}
}
\value{
Valmis esittelytaulukko.
}
\description{
Funktio tekee jatkuvasta muuttujasta HY:n sosiaalitieteiden ohjeen mukaisen esittelytaulukon.
}
