% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_korrelaatiotaulukko}
\alias{tee_korrelaatiotaulukko}
\title{Korrelaatiotaulukko}
\usage{
tee_korrelaatiotaulukko(korrelaatiomatriisi)
}
\arguments{
\item{korrelaatiomatriisi}{psych::corr.test()-objekti tai lista, jossa komponentit r ja p.}
}
\value{
Tibble: Muuttuja + yläkolmion korrelaatiot tähdillä.
}
\description{
Tekee yläkolmion korrelaatiotaulukon, jossa korrelaatiot on pyöristetty ja
p-arvot esitetty merkitsevyystähdillä.
}
