% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_korrelaatiotaulukko_kuvailevin_tunnusluvuin}
\alias{tee_korrelaatiotaulukko_kuvailevin_tunnusluvuin}
\title{Korrelaatiotaulukko kuvailevin tunnusluvuin}
\usage{
tee_korrelaatiotaulukko_kuvailevin_tunnusluvuin(korrelaatiomatriisi, aineisto)
}
\arguments{
\item{korrelaatiomatriisi}{psych::corr.test()-objekti tai lista, jossa komponentit r ja p.}

\item{aineisto}{data.frame, josta lasketaan n, ka ja kh. Sen tulee sisältää samat muuttujat kuin korrelaatiomatriisissa.}
}
\value{
Tibble: Muuttuja, n, ka, kh + yläkolmion korrelaatiot tähdillä.
}
\description{
Tekee yläkolmion korrelaatiotaulukon ja lisää muuttujakohtaiset kuvailevat tunnusluvut:
n, keskiarvo (ka) ja keskihajonta (kh).
}
