% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_log_regressiotaulukko}
\alias{tee_log_regressiotaulukko}
\title{Tehdään logistisen regression taulukko.}
\usage{
tee_log_regressiotaulukko(
  malli,
  luottamustaso = 0.95,
  pyorista_est = 2,
  pyorista_p = 3,
  kieli = "suomi",
  desimaalierotin = "piste"
)
}
\arguments{
\item{malli}{Logistinen regressioanalyysimalli.}

\item{luottamustaso}{Luottamusvälin taso (oletus 0.95).}

\item{pyorista_est}{Desimaalit estimaateille, SE:lle ja z:lle (oletus 2).}

\item{pyorista_p}{Desimaalit p-arvoille (oletus 3; käytetään \code{p_coding}).}

\item{kieli}{Character. \code{"suomi"} (oletus) tai \code{"eng"} englanniksi.}

\item{desimaalierotin}{Jos \code{"pilkku"}, muuttaa desimaalierottimen pilkuksi. Oletus \code{"piste"}.}
}
\value{
Valmis esittelytaulukko.
}
\description{
Funktio tekee logisesta regressiomallista HY:n sosiaalitieteiden ohjeen mukaisen tulostaulukon
}
