% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_luokitellun_muuttujan_esittely}
\alias{tee_luokitellun_muuttujan_esittely}
\title{Esittele luokiteltua muuttujaa}
\usage{
tee_luokitellun_muuttujan_esittely(
  aineisto,
  muuttuja,
  desimaalipaikkoja = 2,
  kieli = "suomi",
  desimaalierotin = "piste"
)
}
\arguments{
\item{aineisto}{Data frame, josta muuttuja lasketaan.}

\item{muuttuja}{Luokiteltava muuttuja (tidy eval -syntaksi).}

\item{desimaalipaikkoja}{Desimaalipaikat prosenttiluvuille (oletus 2).}

\item{kieli}{Character. \code{"suomi"} (oletus) tai \code{"eng"} englanniksi.}

\item{desimaalierotin}{Jos \code{"pilkku"}, muuttaa desimaalierottimen pilkuksi. Oletus \code{"piste"}.}
}
\value{
Tibble: muuttuja, \code{n}, \code{\%} ja tarvittaessa
  \code{\% vastanneista} (kun aineistossa on puuttuvia havaintoja).
}
\description{
Laskee luokitellulle muuttujalle frekvenssit ja prosenttiosuudet
\code{janitor::tabyl()}-funktion avulla.
}
