% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_r2_yhteenveto}
\alias{tee_r2_yhteenveto}
\title{Yhteenvetoteksti: R^2, korjattu R^2 ja ΔR^2 merkitsevyys hierarkkisille lm-malleille}
\usage{
tee_r2_yhteenveto(mallit, digits = 3, p_digits = 3)
}
\arguments{
\item{mallit}{Lista peräkkäin sisäkkäisiä lm-malleja (askel 1, 2, 3, ...).}

\item{digits}{Desimaalit R^2-lukuihin (oletus 3).}

\item{p_digits}{Desimaalit p-arvoihin (oletus 3, käytetään p_coding()).}
}
\value{
Merkkijono, esim:
  "Huom. R2 korj. = .007 askeleella 1., R2 korj. = .018 askeleella 2., R2 korj. = .081 askeleella 3.;
   R2 = .018 askeleella 2 (p = .112), ΔR2 = .069 askeleella 3 (p = .002)"
}
\description{
Yhteenvetoteksti: R^2, korjattu R^2 ja ΔR^2 merkitsevyys hierarkkisille lm-malleille
}
