% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_regressiotaulukko_selitettava_rivina}
\alias{tee_regressiotaulukko_selitettava_rivina}
\title{Regressioanalyysin tulostaulukko, selitettävä rivinä.}
\usage{
tee_regressiotaulukko_selitettava_rivina(malli, kieli = "suomi", desimaalierotin = "piste")
}
\arguments{
\item{malli}{Lineaarinen regressioanalyysimalli.}

\item{kieli}{Character. \code{"suomi"} (oletus) tai \code{"eng"} englanniksi.}

\item{desimaalierotin}{Jos \code{"pilkku"}, muuttaa desimaalierottimen pilkuksi. Oletus \code{"piste"}.}
}
\value{
Valmis regressioanalyysitaulukko.
}
\description{
Funktio tekee lineaarisesta regressioanalyysimallista HY:n sosiaalitieteiden koulutusohjelmien ohjeiden mukaisen tulostaulukon, joss
selitettävä muuttuja rivinä.
}
