% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_korrelaatiotaulukko}
\alias{tee_korrelaatiotaulukko}
\title{Korrelaatiotaulukko}
\usage{
tee_korrelaatiotaulukko(korrelaatiomatriisi)
}
\arguments{
\item{korrelaatiomatriisi}{Korrelaatiomatriisi}
}
\value{
Valmis korrelaatiotaulukko.
}
\description{
Funktio tekee korrelaatiomatriisista HY:n sosiaalitieteiden koulutusohjelmien ohjeiden mukaisen tulostaulukon.
}
