% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_sosiaalitieteiden_taulukointikomennot.R
\name{tee_regressiotaulukko}
\alias{tee_regressiotaulukko}
\title{Regressioanalyysin tulostaulukko}
\usage{
tee_regressiotaulukko(malli, kieli = "suomi")
}
\arguments{
\item{malli}{Lineaarinen regressioanalyysimalli.}

\item{kieli}{Character. \code{"suomi"} (oletus) tai \code{"eng"} englanniksi.}
}
\value{
Valmis regressioanalyysitaulukko.
}
\description{
Funktio tekee lineaarisesta regressioanalyysimallista HY:n sosiaalitieteiden koulutusohjelmien ohjeiden mukaisen tulostaulukon.
}
